% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project_dir.R
\name{create_project_dir}
\alias{create_project_dir}
\title{crear directorio del proyecto}
\usage{
create_project_dir(path, name)
}
\description{
esta funcion cre la esrtrucutra de carpetas para el proyecto.
}
