% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_project_scripts.R
\name{template_project_scripts.}
\alias{template_project_scripts.}
\alias{template_project_scripts}
\title{crear templates de scripts para proyectos de ciencia de datos}
\usage{
template_project_scripts(project_path)
}
\description{
esta funcion crea scripts de ejemplo para proyectos de ciencia de datos.
}
